[lldb] Move DownloadObjectAndSymbolFile to SymbolLocator plugin
This builds on top of the work started in c3a302d to convert
LocateSymbolFile to a SymbolLocator plugin. This commit moves
DownloadObjectAndSymbolFile.
